move Rust to own dir
