+++ packages/quicktype-core/src/language/Rust/index.ts
@@ -1,0 +1,2 @@
+export { RustTargetLanguage, rustOptions } from "./language";
+export { RustRenderer } from "./RustRenderer";
